+++ test/helpers/steal-funds.ts
